+++ ingestors/imap/__init__.py
@@ -1,0 +1,11 @@
+"""
+ingestors/imap package
+"""
+
+__all__ = [
+    "imap_client",
+    "poller",
+    "parser",
+    "repository",
+    "models",
+]